Make the code clearer
